fixing a compatibility bug
Fixing the bug as mentioned in https://github.com/chentingpc/gfn/issues/4: " tensors used as indices must be long or byte tensors when used in latest version of torch_geometric version 1.3.2"

+++ train_eval.py
@@ -171,8 +171,8 @@ def k_fold(dataset, folds, epoch_select):
 
     for i in range(folds):
         train_mask = torch.ones(len(dataset), dtype=torch.uint8)
-        train_mask[test_indices[i]] = 0
-        train_mask[val_indices[i]] = 0
+        train_mask[test_indices[i].long()] = 0
+        train_mask[val_indices[i].long()] = 0
         train_indices.append(train_mask.nonzero().view(-1))
 
     return train_indices, test_indices, val_indices
